server/pyproject.toml
[tool.poetry]
name = "lionelchallenge"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.2.0"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
psycopg2 = "^2.9.9"
drf-yasg = "^1.21.7"
gunicorn = "^22.0.0"
django-cors-headers = "^4.4.0"
drf-nested-routers = "^0.94.1"
django-storages = "^1.14.2"
boto3 = "^1.34.4"


[tool.poetry.group.dev.dependencies]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
